requirements.txt
streamlit 
scikit-learn 
google-genai
numpy
python-dotenv
pdfminer.six